+++ docs/RELEASE-NOTES.md
@@ -1,5 +1,16 @@
 # Release Notes
 
+## 1.6.0 - ?
+
+Breaking changes:
+- (None)
+
+New features:
+- Added `DecimalExtensions.RemoveTrailingZeros` method.
+
+Bug fixes / internal changes:
+- (None)
+
 ## 1.5.0 - 08 July 2022
 
 Breaking changes: